Do Windows-only thing only on Windows
In Windows, there is a distinction between text mode and binary mode stdio. This program needs stdin in binary mode, so on Windows the program needs to set stdin to binary mode before using it. On Linux, there is no distinction between text and binary, so the Linux version of the program just doesn't need to do this.